# docs/openapi.json
{
	"components": {
		"schemas": {
...
			"handlers.ErrorResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"error",
					"message"
				],
				"type": "object"
			},
...
			"handlers.PaginatedInAppNotificationsResponse": {
				"properties": {
					"limit": {
						"type": "integer"
					},
					"notifications": {
						"items": {
							"$ref": "#/components/schemas/models.UserNotification"
						},
						"type": "array"
					},
					"page": {
						"type": "integer"
					},
					"total": {
						"type": "integer"
					},
					"total_pages": {
						"type": "integer"
					}
				},
				"required": [
					"limit",
					"notifications",
					"page",
					"total",
					"total_pages"
				],
				"type": "object"
			},
...
			"handlers.SuccessResponse": {
				"properties": {
					"data": {},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"message"
				],
				"type": "object"
			},
			"handlers.UnreadCountResponse": {
				"properties": {
					"count": {
						"type": "integer"
					}
				},
				"required": [
					"count"
				],
				"type": "object"
			},
...
			"models.UserNotification": {
				"properties": {
					"created_at": {
						"type": "string"
					},
					"id": {
						"type": "integer"
					},
					"payload": {
						"type": "string"
					},
					"read": {
						"type": "boolean"
					},
					"type": {
						"$ref": "#/components/schemas/models.UserNotificationType"
					},
					"updated_at": {
						"type": "string"
					},
					"user_id": {
						"type": "integer"
					}
				},
				"required": [
					"created_at",
					"id",
					"payload",
					"read",
					"type",
					"updated_at",
					"user_id"
				],
				"type": "object"
			},
			"models.UserNotificationType": {
				"enum": [
					"group_invite"
				],
				"type": "string",
				"x-enum-varnames": [
					"UserNotificationTypeGroupInvite"
				]
			},
...
		},
		"securitySchemes": {
			"BearerAuth": {
				"description": "Type \"Bearer\" followed by a space and JWT token.",
				"in": "header",
				"name": "Authorization",
				"type": "apiKey"
			}
		}
	},
	"info": {
		"contact": {
			"email": "support@example.com",
			"name": "API Support"
		},
		"description": "A RESTful API for managing tasks with JWT authentication and user task assignment",
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		},
		"termsOfService": "http://swagger.io/terms/",
		"title": "Todo API",
		"version": "1.0"
	},
	"openapi": "3.0.3",
	"paths": {
...
		"/notifications/in-app": {
			"get": {
				"description": "Returns in-app notifications (e.g. group invites). Use unread_only=true for unread only.",
				"parameters": [
					{
						"description": "Page number (default: 1)",
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "Items per page (default: 20, max: 100)",
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "Only unread notifications",
						"in": "query",
						"name": "unread_only",
						"schema": {
							"type": "boolean"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.PaginatedInAppNotificationsResponse"
								}
							}
						},
						"description": "OK"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Unauthorized"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				],
				"summary": "List in-app notifications",
				"tags": [
					"notifications-in-app"
				]
			}
		},
		"/notifications/in-app/read-all": {
			"post": {
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.SuccessResponse"
								}
							}
						},
						"description": "OK"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Unauthorized"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				],
				"summary": "Mark all notifications as read",
				"tags": [
					"notifications-in-app"
				]
			}
		},
		"/notifications/in-app/unread-count": {
			"get": {
				"description": "Returns how many in-app notifications are unread (for badge display).",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.UnreadCountResponse"
								}
							}
						},
						"description": "OK"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Unauthorized"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				],
				"summary": "Unread in-app notification count",
				"tags": [
					"notifications-in-app"
				]
			}
		},
		"/notifications/in-app/{id}/read": {
			"patch": {
				"parameters": [
					{
						"description": "Notification ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.SuccessResponse"
								}
							}
						},
						"description": "OK"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Unauthorized"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Forbidden"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				],
				"summary": "Mark notification as read",
				"tags": [
					"notifications-in-app"
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://localhost:8080/api/v1"
		}
	]
}